fix: ajustando campos e removendo itens desnecessários
