+++ .env.example
@@ -1,1 +1,2 @@
 DATABASE_URL="postgresql://postgres:postgres@localhost:5432/mydb"
+JWT_SECRET="{SECRET_SERVER_LUCAO}"
