+++ src/app.module.ts
@@ -3,13 +3,15 @@ import { ConfigModule } from '@nestjs/config';
 import { AuthModule } from './auth/auth.module';
 import { AlunoModule } from './modules/aluno/aluno.module';
 import { ProfessorModule } from './modules/professor/professor.module';
+import { SolicitacaoDefesaModule } from './modules/solicitacao-defesa/solicitacao-defesa.module';
 
 @Module({
   imports: [
     ConfigModule.forRoot({ isGlobal: true }),
+    AuthModule,
     AlunoModule,
     ProfessorModule,
-    AuthModule,
+    SolicitacaoDefesaModule,
   ],
 })
 export class AppModule {}
