+++ src/modules/solicitacao-defesa/dto/update-solicitacao-defesa.dto..ts
@@ -1,0 +1,4 @@
+import { PartialType } from "@nestjs/mapped-types";
+import { CreateSolicitacaoDefesaDto } from "./create-solicitacao-defesa.dto";
+
+export class UpdateSolicitacaoDefesa extends PartialType(CreateSolicitacaoDefesaDto){}
