+++ src/modules/solicitacao-defesa/solicitacao-defesa.module.ts
@@ -1,9 +1,10 @@
 import { Module } from '@nestjs/common';
+import { PrismaService } from 'src/service/prisma.service';
 import { SolicitacaoDefesaController } from './solicitacao-defesa.controller';
 import { SolicitacaoDefesaService } from './solicitacao-defesa.service';
 
 @Module({
   controllers: [SolicitacaoDefesaController],
-  providers: [SolicitacaoDefesaService]
+  providers: [SolicitacaoDefesaService, PrismaService]
 })
 export class SolicitacaoDefesaModule {}
